Fix bet-path overflow, cap interest, disable late segment removal
- Bet placement path int -> long (Bet.amount, currentlyInHand, Chip,
  totalAmount, all createBet/makeChip regions, chip-click Optional). Fixes
  the chip-value int overflow at balances >2.1B (garbage labels / negative
  bets seen at $4.4B in a stage-7 overshoot run). Range now ~9.2e18.
- Interest percentage now taken on min(balance, stage goal), so being
  richer than the goal earns no extra interest -> it can no longer
  passively clear a stage or win a boss gain-goal by doing nothing.
- Segment Remover disabled from stage 4 on (DELETE_BUDGET {5,3,1,0}); it's
  now a strictly early-game odds-shaping tool.

+++ core/src/main/java/de/mario/roguelette/Player.java
@@ -10,7 +10,7 @@ public class Player {
     private final Inventory inventory;
 
     private long balance;
-    private int currentlyInHand = 0; //amount of money currently in hand to be placed somewhere
+    private long currentlyInHand = 0; //amount of money currently in hand to be placed somewhere
     private final String name;
 
     private final Character character;
@@ -40,11 +40,11 @@ public class Player {
         return balance;
     }
 
-    public int getCurrentlyInHand() {
+    public long getCurrentlyInHand() {
         return currentlyInHand;
     }
 
-    public void increaseHandBy(int amount) {
+    public void increaseHandBy(long amount) {
         currentlyInHand += amount;
     }
